+++ spec/command/server_spec.rb
@@ -3,11 +3,23 @@ require 'spec_helper'
 describe Miron::Command::Server do
   include SpecHelper::Command
 
-  before(:each) do
-    create_mironfile(SpecHelper.temporary_directory)
-  end
+  describe 'CLI Parameters' do
+    before(:each) do
+      create_mironfile(SpecHelper.temporary_directory)
+    end
 
-  context 'mironfile exists' do
+    it 'accepts --handler as a port argument' do
+      Dir.chdir(SpecHelper.temporary_directory) do
+        expect { run_command('server', '--handler=webrick') }.to_not raise_error(CLAide::Help)
+      end
+    end
+
+    it 'accepts --mironfile as a port argument' do
+      Dir.chdir(SpecHelper.temporary_directory) do
+        expect { run_command('server', '--mironfile=Mironfile.rb') }.to_not raise_error(CLAide::Help)
+      end
+    end
+
     it 'accepts --port as a port argument' do
       Dir.chdir(SpecHelper.temporary_directory) do
         expect { run_command('server', '--port=9290') }.to_not raise_error(CLAide::Help)
@@ -15,11 +27,22 @@ describe Miron::Command::Server do
     end
   end
 
-  context 'mironfile does not exist' do
-    it 'complains if no mironfile is present' do
-      Dir.chdir(SpecHelper.temporary_directory) do
+  describe 'Mironfile handling' do
+    context 'mironfile does not exist' do
+      before do
         remove_mironfile(SpecHelper.temporary_directory)
-        expect { run_command('server', '--port=9290') }.to raise_error(Errno::ENOENT)
+      end
+
+      it 'complains if no mironfile is present (with no parameter)' do
+        Dir.chdir(SpecHelper.temporary_directory) do
+          expect { run_command('server') }.to raise_error(CLAide::Help)
+        end
+      end
+
+      it 'complains if no mironfile is present (with parameter)' do
+        Dir.chdir(SpecHelper.temporary_directory) do
+          expect { run_command('server', '--mironfile=lalala.rb') }.to raise_error(CLAide::Help)
+        end
       end
     end
   end