+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Next Release
-* Your contribution here.
+* Pass paramters to middleware and app - [@maclover7](https://github.com/maclover7).
 
 ### 0.0.2 - (10/8/2015)
 * Add `Miron::Auth::Basic` middleware to allow HTTP Basic authentication - [@maclover7](https://github.com/maclover7).