+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Redo Miron::Auth::Basic to accept params via Mironfile. Closes #15. - [@maclover7](https://github.com/maclover7).
 * Put all Miron middleware under the `Miron::Middleware` module. - [@maclover7](https://github.com/maclover7).
 * Add `Miron::Middleware::Static` to serve static files. Closes #16. - [@maclover7](https://github.com/maclover7).
+* Allow specification of a custom file (not just Mironfile.rb) in `miron
+  server` command - [@maclover7](https://github.com/maclover7).
 
 ### 0.0.2 - (10/8/2015)
 * Add `Miron::Auth::Basic` middleware to allow HTTP Basic authentication - [@maclover7](https://github.com/maclover7).