+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### Next Release
+### 0.0.3 - (10/14/2015)
 * Pass paramters to middleware and app - [@maclover7](https://github.com/maclover7).
 * Redo Miron::Auth::Basic to accept params via Mironfile. Closes #15. - [@maclover7](https://github.com/maclover7).
 * Put all Miron middleware under the `Miron::Middleware` module. - [@maclover7](https://github.com/maclover7).